Update help component after removing bootstrap
